--- pkg/plugins/helm/main.go
@@ -1,1 +1,0 @@
-package helm
