+++ pkg/plugins/helmchart/main_test.go
@@ -1,4 +1,4 @@
-package chart
+package helmchart
 
 import "testing"
